+++ build-scripts/build-client-scripts.js
