+++ build-scripts/do-client-build.js
