+++ build-scripts/generate-vite-config-for-entry-point.js
