+++ build-scripts/write-version-files.js
